Build path
Configuration of the virtual environment and generation of the "requirements.txt" file which holds the dependencies of the project. VENV is independent of path, ide, and editor. In other words this isolates our projects dependancies and provides means for a user to install all dependancies and run with minimal effort