Entire bingo implementation
There is a lot here.
The functions created are:
bingoHint(required, letters): This is the main bingo hint driver and the only call that is required (all other functions are called in this one)
populateBingo(bingoSheet, required, letters): This function drives the population of the bingo sheet (filling in the values in the appropriate spots)
popHelper(row, starter, required, letters): This function assists the populateBingo function by finding the values which are placed in the bingo sheet
populateSigma(bingo): This function is the final step to complete the bingo hint. It is responsible for calculating and placing all Sigma values

I have tested the functionality of this code in an alternate environment and provided a sample output in the discord. The ultimate output of this group of functions is a 2 dimensional array in the format specified by customer