TAB COMPLETION COMPLETE!
A little hard to follow the changes here so Ill do my best:
added a class to the top of viewCLI.py - this class is responsible for the tab completion and it contains an initializer and a single function:
basically for every command option, if the option starts with what is entered in the command line and u hit tab it will auto complete and u can then press enter

This class is called in the mainGameDisplay(puzzle) function which outputs the puzzle info, command options and prompts user to enter a guess. One change I made here was adding a return statement to this function. This may not be ideal but I really don't see a way around it. I made the subsequent change to the mainGameCLI call to this function and I believe it to be effective (unless there is another call to mainGameDispaly() that I missed which I don thtink there is). I ran the program several times and it had no issues for me. Additionally I updated the requirements.txt file to include the library I used to accomplish tab completion