Encryption
used crpytography library to encrpyt. Functions for encryption are in model class, function returns an encrypted list of valid words. updated requirements.txt and it worked fine
all other changes are for the cli view during saving and loading. if the file we are loading does not have an author filed set to "BigPythonEnergy" than the "decryption" fails and restarts the start screen. Otherwise the file is "successfully decrypted" and the load the game as per usual
during saving after u enter to save a prompt will be displayed asking weither to encrypt, if u say yes than it will encrpyt otherwise is saves the field noramlly